refactor: extract guards
