+++ packages/pure-parse/src/guard/README.test.ts
@@ -1,6 +1,10 @@
 import { describe, expect, it } from 'vitest'
-import { arrayGuard, literal, objectGuard, union, Guard } from './validation'
+import { arrayGuard } from './validation'
 import { isString } from './primitives'
+import { Guard } from './types'
+import { literal } from './literal'
+import { union } from './union'
+import { objectGuard } from './object'
 
 describe('README examples', () => {
   describe('Tree', () => {
