+++ packages/pure-parse/src/guard/index.ts
@@ -1,3 +1,9 @@
 export * from './validation'
 export * from './primitives'
 export * from './json'
+export * from './types'
+export * from './literal'
+export { union } from './union'
+export { tuple } from './tuple'
+export { objectGuard } from './object'
+export { objectGuardNoJit } from './object'
