+++ packages/pure-parse/src/guard/json.ts
@@ -1,6 +1,7 @@
-import { arrayGuard, partialRecord, union } from './validation'
+import { arrayGuard, partialRecord } from './validation'
 import { isBoolean, isNull, isNumber, isString } from './primitives'
 import { JsonValue } from '../common'
+import { union } from './union'
 
 export const isJsonValue = (data: unknown): data is JsonValue =>
   union(
