+++ packages/pure-parse/src/guard/tuple.ts
@@ -1,0 +1,17 @@
+import { Guard } from './types'
+
+/**
+ * @param guards an array of guards. Each guard validates the corresponding element in the data tuple.
+ */
+export const tuple =
+  <T extends readonly [...unknown[]]>(
+    guards: [
+      ...{
+        [K in keyof T]: Guard<T[K]>
+      },
+    ],
+  ): Guard<T> =>
+  (data: unknown): data is T =>
+    Array.isArray(data) &&
+    data.length === guards.length &&
+    guards.every((guard, index) => guard(data[index]))
