+++ packages/pure-parse/src/guard/types.ts
@@ -1,0 +1,12 @@
+import { optionalSymbol } from '../internals'
+
+/**
+ * A function that returns a [type predicate](https://www.typescriptlang.org/docs/handbook/advanced-types.html#using-type-predicates) on the argument.
+ */
+export type Guard<T> = (data: unknown) => data is T
+/**
+ * Special guard to check optional values
+ */
+export type OptionalGuard<T> = { [optionalSymbol]: true } & ((
+  data: unknown,
+) => data is typeof optionalSymbol | T | undefined)
