+++ packages/pure-parse/src/parse/index.ts
@@ -5,3 +5,4 @@ export * from './union'
 export * from './primitives'
 export * from './fallback'
 export * from './json'
+export { literal } from './literal'
