+++ packages/pure-parse/src/parse/primitives.test.ts
@@ -1,6 +1,6 @@
 import { describe, expect, it, test } from 'vitest'
-import { literal } from './primitives'
 import { fallback } from './fallback'
+import { literal } from './literal'
 
 describe('primitives', () => {
   describe.todo('null')
@@ -10,20 +10,4 @@ describe('primitives', () => {
   describe.todo('bigint')
   describe.todo('string')
   describe.todo('symbol')
-  describe('literals', () => {
-    it('todo', () => {
-      const parseLiteral = literal('a')
-      expect(parseLiteral('a')).toEqual({
-        tag: 'success',
-        value: 'a',
-      })
-    })
-    test('with fallback', () => {
-      const parseLiteral = fallback(literal('#FF0000'), '#00FF00')
-      expect(parseLiteral('#XXYYZZ')).toEqual({
-        tag: 'success',
-        value: '#00FF00',
-      })
-    })
-  })
 })
